requirements.txt
Flask==3.0.3
python-socketio==5.11.3
SQLAlchemy==2.0.32
